projekt_studia/requirements.txt
Django==5.1.5
psycopg[binary]==3.1.18
pillow==10.2.0
